requirements.txt
Django~=4.0.4
djangorestframework~=3.13.1
django-debug-toolbar==3.2.4
drf-yasg==1.20.0
flake8==4.0.1
black==22.3.0
Pillow==9.1.0
requests==2.27.1
django_filters==21.1